+++ rust/kafka/src/batch.rs
@@ -1,15 +1,32 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::sync::Arc;
 
 use rdkafka::message::BorrowedMessage;
 use rdkafka::Message;
 
-use arrow::error::Result as ArrowResult;
-use arrow::datatypes::{DataType, Field};
 use arrow::array::{
     Array, BinaryBuilder, Int32Builder, Int64Builder, StringBuilder, StructArray,
 };
+use arrow::datatypes::{DataType, Field};
+use arrow::error::Result as ArrowResult;
 
-pub (crate) struct KafkaBatch {
+pub(crate) struct KafkaBatch {
     nread: usize,
     topics: StringBuilder,
     offsets: Int64Builder,
@@ -19,7 +36,7 @@ pub (crate) struct KafkaBatch {
 }
 
 impl KafkaBatch {
-    pub (crate) fn new(max_batch_size: usize) -> Self {
+    pub(crate) fn new(max_batch_size: usize) -> Self {
         let topics = StringBuilder::new(max_batch_size);
         let offsets = Int64Builder::new(max_batch_size);
         let partitions = Int32Builder::new(max_batch_size);
@@ -31,15 +48,16 @@ impl KafkaBatch {
             offsets,
             partitions,
             keys,
-            payloads
+            payloads,
         }
     }
 
-    pub (crate) fn len(&self) -> usize {
+    pub(crate) fn len(&self) -> usize {
         self.nread
     }
 
-    pub (crate) fn process_message(&mut self,
+    pub(crate) fn process_message(
+        &mut self,
         message: &BorrowedMessage,
     ) -> ArrowResult<()> {
         self.nread += 1;
@@ -86,5 +104,3 @@ impl From<KafkaBatch> for StructArray {
         fields.into()
     }
 }
-
-
